Split large test class
